split test_multi.py into separate files and address the comments from pull request
