+++ pandas/tests/indexes/multi/test_analytics.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+def test_shift(idx):
+
+    # GH8083 test the base class for shift
+    pytest.raises(NotImplementedError, idx.shift, 1)
+    pytest.raises(NotImplementedError, idx.shift, 1, 2)
