+++ pandas/tests/indexes/multi/test_constructor.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 import pandas.util.testing as tm
 import pytest
-from pandas import DataFrame, Index, MultiIndex, date_range
+from pandas import Index, MultiIndex, date_range
 from pandas._libs.tslib import Timestamp
 from pandas.compat import lrange, range
 from pandas.core.dtypes.cast import construct_1d_object_array_from_listlike
@@ -89,82 +89,6 @@ def test_copy_in_constructor():
     val = levels[0]
     levels[0] = "PANDA"
     assert mi.levels[0][0] == val
-
-
-def test_reconstruct_sort():
-
-    # starts off lexsorted & monotonic
-    mi = MultiIndex.from_arrays([
-        ['A', 'A', 'B', 'B', 'B'], [1, 2, 1, 2, 3]
-    ])
-    assert mi.is_lexsorted()
-    assert mi.is_monotonic
-
-    recons = mi._sort_levels_monotonic()
-    assert recons.is_lexsorted()
-    assert recons.is_monotonic
-    assert mi is recons
-
-    assert mi.equals(recons)
-    assert Index(mi.values).equals(Index(recons.values))
-
-    # cannot convert to lexsorted
-    mi = pd.MultiIndex.from_tuples([('z', 'a'), ('x', 'a'), ('y', 'b'),
-                                    ('x', 'b'), ('y', 'a'), ('z', 'b')],
-                                   names=['one', 'two'])
-    assert not mi.is_lexsorted()
-    assert not mi.is_monotonic
-
-    recons = mi._sort_levels_monotonic()
-    assert not recons.is_lexsorted()
-    assert not recons.is_monotonic
-
-    assert mi.equals(recons)
-    assert Index(mi.values).equals(Index(recons.values))
-
-    # cannot convert to lexsorted
-    mi = MultiIndex(levels=[['b', 'd', 'a'], [1, 2, 3]],
-                    labels=[[0, 1, 0, 2], [2, 0, 0, 1]],
-                    names=['col1', 'col2'])
-    assert not mi.is_lexsorted()
-    assert not mi.is_monotonic
-
-    recons = mi._sort_levels_monotonic()
-    assert not recons.is_lexsorted()
-    assert not recons.is_monotonic
-
-    assert mi.equals(recons)
-    assert Index(mi.values).equals(Index(recons.values))
-
-
-def test_reconstruct_remove_unused():
-    # xref to GH 2770
-    df = DataFrame([['deleteMe', 1, 9],
-                    ['keepMe', 2, 9],
-                    ['keepMeToo', 3, 9]],
-                   columns=['first', 'second', 'third'])
-    df2 = df.set_index(['first', 'second'], drop=False)
-    df2 = df2[df2['first'] != 'deleteMe']
-
-    # removed levels are there
-    expected = MultiIndex(levels=[['deleteMe', 'keepMe', 'keepMeToo'],
-                                  [1, 2, 3]],
-                          labels=[[1, 2], [1, 2]],
-                          names=['first', 'second'])
-    result = df2.index
-    tm.assert_index_equal(result, expected)
-
-    expected = MultiIndex(levels=[['keepMe', 'keepMeToo'],
-                                  [2, 3]],
-                          labels=[[0, 1], [0, 1]],
-                          names=['first', 'second'])
-    result = df2.index.remove_unused_levels()
-    tm.assert_index_equal(result, expected)
-
-    # idempotent
-    result2 = result.remove_unused_levels()
-    tm.assert_index_equal(result2, expected)
-    assert result2.is_(result)
 
 
 def test_from_arrays(idx):
