+++ pandas/tests/indexes/multi/test_set_ops.py
@@ -150,13 +150,6 @@ def test_empty(idx):
     assert idx[:0].empty
 
 
-def test_unique_na():
-    idx = pd.Index([2, np.nan, 2, 1], name='my_index')
-    expected = pd.Index([2, np.nan, 1], name='my_index')
-    result = idx.unique()
-    tm.assert_index_equal(result, expected)
-
-
 def test_difference(idx):
 
     first = idx
