+++ pandas/tests/indexes/multi/test_sorting.py
@@ -3,7 +3,7 @@ import numpy as np
 import pandas as pd
 import pandas.util.testing as tm
 import pytest
-from pandas import CategoricalIndex, DataFrame, MultiIndex, RangeIndex
+from pandas import CategoricalIndex, DataFrame, Index, MultiIndex, RangeIndex
 from pandas.compat import lrange
 from pandas.errors import PerformanceWarning, UnsortedIndexError
 
@@ -135,3 +135,122 @@ def test_unsortedindex_doc_examples():
 
     assert dfm.index.is_lexsorted()
     assert dfm.index.lexsort_depth == 2
+
+
+def test_reconstruct_sort():
+
+    # starts off lexsorted & monotonic
+    mi = MultiIndex.from_arrays([
+        ['A', 'A', 'B', 'B', 'B'], [1, 2, 1, 2, 3]
+    ])
+    assert mi.is_lexsorted()
+    assert mi.is_monotonic
+
+    recons = mi._sort_levels_monotonic()
+    assert recons.is_lexsorted()
+    assert recons.is_monotonic
+    assert mi is recons
+
+    assert mi.equals(recons)
+    assert Index(mi.values).equals(Index(recons.values))
+
+    # cannot convert to lexsorted
+    mi = pd.MultiIndex.from_tuples([('z', 'a'), ('x', 'a'), ('y', 'b'),
+                                    ('x', 'b'), ('y', 'a'), ('z', 'b')],
+                                   names=['one', 'two'])
+    assert not mi.is_lexsorted()
+    assert not mi.is_monotonic
+
+    recons = mi._sort_levels_monotonic()
+    assert not recons.is_lexsorted()
+    assert not recons.is_monotonic
+
+    assert mi.equals(recons)
+    assert Index(mi.values).equals(Index(recons.values))
+
+    # cannot convert to lexsorted
+    mi = MultiIndex(levels=[['b', 'd', 'a'], [1, 2, 3]],
+                    labels=[[0, 1, 0, 2], [2, 0, 0, 1]],
+                    names=['col1', 'col2'])
+    assert not mi.is_lexsorted()
+    assert not mi.is_monotonic
+
+    recons = mi._sort_levels_monotonic()
+    assert not recons.is_lexsorted()
+    assert not recons.is_monotonic
+
+    assert mi.equals(recons)
+    assert Index(mi.values).equals(Index(recons.values))
+
+
+def test_reconstruct_remove_unused():
+    # xref to GH 2770
+    df = DataFrame([['deleteMe', 1, 9],
+                    ['keepMe', 2, 9],
+                    ['keepMeToo', 3, 9]],
+                   columns=['first', 'second', 'third'])
+    df2 = df.set_index(['first', 'second'], drop=False)
+    df2 = df2[df2['first'] != 'deleteMe']
+
+    # removed levels are there
+    expected = MultiIndex(levels=[['deleteMe', 'keepMe', 'keepMeToo'],
+                                  [1, 2, 3]],
+                          labels=[[1, 2], [1, 2]],
+                          names=['first', 'second'])
+    result = df2.index
+    tm.assert_index_equal(result, expected)
+
+    expected = MultiIndex(levels=[['keepMe', 'keepMeToo'],
+                                  [2, 3]],
+                          labels=[[0, 1], [0, 1]],
+                          names=['first', 'second'])
+    result = df2.index.remove_unused_levels()
+    tm.assert_index_equal(result, expected)
+
+    # idempotent
+    result2 = result.remove_unused_levels()
+    tm.assert_index_equal(result2, expected)
+    assert result2.is_(result)
+
+
+@pytest.mark.parametrize('first_type,second_type', [
+    ('int64', 'int64'),
+    ('datetime64[D]', 'str')])
+def test_remove_unused_levels_large(first_type, second_type):
+    # GH16556
+
+    # because tests should be deterministic (and this test in particular
+    # checks that levels are removed, which is not the case for every
+    # random input):
+    rng = np.random.RandomState(4)  # seed is arbitrary value that works
+
+    size = 1 << 16
+    df = DataFrame(dict(
+        first=rng.randint(0, 1 << 13, size).astype(first_type),
+        second=rng.randint(0, 1 << 10, size).astype(second_type),
+        third=rng.rand(size)))
+    df = df.groupby(['first', 'second']).sum()
+    df = df[df.third < 0.1]
+
+    result = df.index.remove_unused_levels()
+    assert len(result.levels[0]) < len(df.index.levels[0])
+    assert len(result.levels[1]) < len(df.index.levels[1])
+    assert result.equals(df.index)
+
+    expected = df.reset_index().set_index(['first', 'second']).index
+    tm.assert_index_equal(result, expected)
+
+
+@pytest.mark.parametrize('level0', [['a', 'd', 'b'],
+                                    ['a', 'd', 'b', 'unused']])
+@pytest.mark.parametrize('level1', [['w', 'x', 'y', 'z'],
+                                    ['w', 'x', 'y', 'z', 'unused']])
+def test_remove_unused_nan(level0, level1):
+    # GH 18417
+    mi = pd.MultiIndex(levels=[level0, level1],
+                       labels=[[0, 2, -1, 1, -1], [0, 1, 2, 3, 2]])
+
+    result = mi.remove_unused_levels()
+    tm.assert_index_equal(result, mi)
+    for level in 0, 1:
+        assert('unused' not in result.levels[level])
